parsers/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.9
lxml>=5.0.0
